// src/app/modules/songs/interfaces/artist-response.interface.ts
// Generated by https://quicktype.io

export interface ArtistResponse {
  external_urls: ExternalUrls;
  followers:     Followers;
  genres:        string[];
  href:          string;
  id:            string;
  images:        Image[];
  name:          string;
  popularity:    number;
  type:          string;
  uri:           string;
}

export interface ExternalUrls {
  spotify: string;
}

export interface Followers {
  href:  null;
  total: number;
}

export interface Image {
  height: number;
  url:    string;
  width:  number;
}


// Version short for response
export interface ArtistShort {
  id:            string;
  name:          string;
  images:        Image[];
  external_urls: ExternalUrls;
}