// src/app/modules/songs/interfaces/tracks-response.interface.ts
// Generated by https://quicktype.io

export interface TracksResponse {
  tracks: Track[];
}

export interface Track {
  album:         Album;
  artists:       Artist[];
  disc_number:   number;
  duration_ms:   number;
  explicit:      boolean;
  external_ids:  ExternalIDS;
  external_urls: ExternalUrls;
  href:          string;
  id:            string;
  is_local:      boolean;
  is_playable:   boolean;
  name:          string;
  popularity:    number;
  preview_url:   null | string;
  track_number:  number;
  type:          TrackType;
  uri:           string;
}

export interface Album {
  album_type:             AlbumTypeEnum;
  artists:                Artist[];
  external_urls:          ExternalUrls;
  href:                   string;
  id:                     string;
  images:                 Image[];
  name:                   string;
  release_date:           string;
  release_date_precision: ReleaseDatePrecision;
  total_tracks:           number;
  type:                   AlbumTypeEnum;
  uri:                    string;
}

export enum AlbumTypeEnum {
  Album = "album",
  Single = "single",
}

export interface Artist {
  external_urls: ExternalUrls;
  href:          string;
  id:            string;
  name:          string;
  type:          ArtistType;
  uri:           string;
}

export interface ExternalUrls {
  spotify: string;
}

export enum ArtistType {
  Artist = "artist",
}

export interface Image {
  height: number;
  url:    string;
  width:  number;
}

export enum ReleaseDatePrecision {
  Day = "day",
}

export interface ExternalIDS {
  isrc: string;
}

export enum TrackType {
  Track = "track",
}


// Version short for response
export interface TrackShort {
  id:                     string;
  songName:               string;
  albumName:              string;
  images:                 Image[];
  songURI:                string;
}